Delete Line ".setMetadataAdapter(new JavaReflectionAdapter()));" and move SomeOtherClass.java
and SomeOtherClass2.java to com.querydsl.codegen package.
