+++ querydsl-codegen/src/test/java/com/querydsl/codegen/SomeOtherClass.java
@@ -1,4 +1,4 @@
-package com.querydsl;
+package com.querydsl.codegen;
 
 public class SomeOtherClass {
