+++ querydsl-codegen/src/test/java/com/querydsl/codegen/SomeOtherClass2.java
@@ -1,4 +1,4 @@
-package com.querydsl;
+package com.querydsl.codegen;
 
 public class SomeOtherClass2 {
